fix: keep timer running across navigation
